+++ src/main/java/menu/Application.java
@@ -1,15 +1,14 @@
 package menu;
 
 import menu.controller.MenuController;
-import menu.service.MenuService;
-import menu.service.RecommendService;
+import menu.domain.Recommender;
 
 public class Application {
 
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        MenuController menuController = new MenuController(InitialMenu.initMap(), new MenuService(),
-            new RecommendService());
+        MenuController menuController = new MenuController(InitialMenu.initMap(),
+            new Recommender());
         menuController.start();
     }
 }
